requirements.txt
django==4.2
djangorestframework==3.14.0
drf_spectacular==0.26.2
drf-spectacular-sidecar==2023.4.1
django-mptt==0.14.0
pillow==9.5.0
django-debug-toolbar==4.0.0
python_decouple==3.8
psycopg2-binary==2.9.6
